- move HashingAlgorithm, HashingAlgorithm, MD5HashingAlgorithmTest from logic layer to data layer. - add one test case for white spaces.
