+++ src/main/kotlin/data/security/hashing/HashingAlgorithm.kt
@@ -1,4 +1,4 @@
-package logic.security.hashing
+package data.security.hashing
 
 interface HashingAlgorithm {
     fun hashData(data: String): String
