+++ src/main/kotlin/data/security/hashing/HashingException.kt
@@ -1,4 +1,4 @@
-package logic.security.exceprion
+package data.security.hashing
 
 sealed class HashingException(message: String): Exception(message) {
     class BlankDataException: HashingException("hashing function can't operate on blank data")
